Remove gen-graphs dependency from edge
