+++ gen-core/src/graph.rs
@@ -17,6 +17,28 @@ pub struct GraphEdge {
     pub created_on: i64,
 }
 
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub struct GraphLoadBlock {
+    pub id: i64,
+    pub node_id: HashId,
+    pub start: i64,
+    pub end: i64,
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub struct GraphLoadEdge {
+    pub edge_id: HashId,
+    pub source_node_id: HashId,
+    pub source_coordinate: i64,
+    pub source_strand: Strand,
+    pub target_node_id: HashId,
+    pub target_coordinate: i64,
+    pub target_strand: Strand,
+    pub chromosome_index: i64,
+    pub phased: i64,
+    pub created_on: i64,
+}
+
 /// A contiguous slice of a stored sequence represented as a node in graph space.
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Ord, PartialOrd, Deserialize, Serialize)]
 pub struct GraphNode {
